refactor(cli): reorganize file names
