+++ packages/ui/src/cli/bin.ts
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-import { main } from "./index";
+import { main } from "./main";
 
 const argv = process.argv.slice(2).filter((arg) => arg !== "--");
